EXTENSION: Fix FTE_PEXT_TRANS.
The extension was incorrectly implemented from the beginning. To support
the extra bits support for PF_EXTRA_PFS is needed which results in an
extra byte of player flags. With a 24 bit player flags PF_ONGROUND and
PF_SOLID are then found at a higher offset. This added offset is
typically emulated by upshifting the relevant bits in incoming data if
PEXT_TRANS is not part of the extensions the client announces support
for to be able to use the same definitions in both cases.