EXTENSION: Finish FTE_PEXT_COLOURMOD definitions.
Support for colourmod was partially declared in protocol.h, but missing
the entity_state_t bytes, as well as a few defines.

colourmod[3] is r/g/b with 1/32 precision, thus a value of 32 corresponds
to 1.0f, which in effect means no color modification, and the same is
true for 0.0f. Values above 1.0f yields overbrights.